building: Move stock routing priorities to building struct.
Store the stock routing priority in building instead of flag, and store the requested resource type in building as well. This allows cleaner code for requesting resources and routing these correctly. It was necessary to add a pseudo resource type representing any food type, since this is the way food is requested by the mines. A RESOURCE_GROUP_FOOD was added to resource_type_t.