viewport: Deactivate "fast building" (fixes #49, closes #51).
The "fast building" option interferes with normal double click e.g. when the cursor is on a flag double clicking will start road construction, making it impossible to open the "transport info" box. The settings for "fast building" are still available in the options box but should be removed later if no solution is found.