game-init: Add GUI component for game initialization
This box is show when the game is started, and can be reopened later using F10.
Currently this box allows a new game to be started, either from a random map or from a predefined mission.
The map size or mission level can be chosen for the two game types respectively.