serf: Allow two serfs to change places independently of their individual states
Add function serf_is_waiting() and serf_switch_waiting(). The first checks if a serf is waiting to walk and in which direction; the second signals the serf to actually begin walking in the given direction by switching place with another serf. These functions make it possible for any state to switch with any other state.
Add parameter to get_walking_animation() to get special walking animation for switching with other serf.