Fix bug where uninitialized fields in player object causes crash
This would happen when inactive players were erroneously marked as
active because the allocation of the player object was not zero'ed out
(this particular problem appeared after loading a save game because
only active players are saved/loaded). In the master branch this is
fixed simply by not allocating the player objects until they are
active but this requires substantial code changes. For now we just
calloc the player objects in this branch which should ensure that the
correct player objects are active after loading a save game.