Port to SDL2 and SDL2_mixer
This is just a simple substitution of SDL2 for SDL, no improvements have
been made to take advantage of the new acceleration framework in SDL2.
Everything is still rendered to a screen in main memory and then
transfered to a texture just before being moved to the screen. Ideally
the new texture feature should be used for sprites but that will require
some rework of the sprite caching and rendering, in particular the
landscape tiles.