map-generator: Move map generator function to separate class
Adds interface MapGenerator which is responsible for generating a
new map. A subclass ClassicMapGenerator implements the current
map generator from the original game. A furter subclass,
ClassicMissionMapGenerator, provides a convenience for generating
maps for missions.

This changes map_t to have the tile data initialized from a
MapGenerator. The initialization works in three steps described in
the comment on map_t.