map: Rename hexagon type check method and make range inclusive
Changes the method name to hexagon_types_in_range(), and makes the
range that is checked inclusive with respect to both min and max.
This makes it more clear what type range is checked. Also changes
the two methods that call this, init_objects_shared() and
init_resources_shared() which also now use an inclusive range. The
calls to these functions have been changed to use inclusive
ranges and the methods have been documented.