map-generator: Fix bug where edge objects were incorrectly removed
Fixes a bug introduced in 53365ae and 88e4a6e which caused some
objects on the edges of the map to be removed incorrectly. This
caused mission maps to be slightly different from the original game
(for example, for mission 1 the bug caused to stone objects to be
incorrectly removed at 0,53 and 0,48).