serf: Record output of logging, fishing, stonecutting, farming
When lumberjack, fisher, stonecutter or farmer drops a resource at
the building flag, the output is recorded in the building progress
field the same way as mines record output. If the serf spends too
much time in the planning state, the absense of output is also
recorded. This requires a counter to be added for these states.