Fix regression causing incorrect land owner
Fix regression introduced in 5862cb3 where owner == 0 was changed to
mean that the land has no owner. This resulted in the first player
(player 0) not being able to own any land. Fixed by incrementing
the stored player number by one such that 0 indicates no owner,
player 0 is indicated by 1, and so on.