game: Fix regression causing failed loading of save games
Fix bug which seems to have been introduced in 3eab668 and
caused a save game without any buildings to fail loading. The
loader would raise a "Data integrity is broken" error. The
cause of this was too many calls to allocate_objects() in
Game. This method allocates dummy (zero index) objects for
serf, flag and building collections but this method was
called both in the Game constructor and during loading of the
game resulting in more than one dummy object being created.
This was fixed by only calling allocate_objects() in the
constructor.