Move total gold counter to Game
Previously the Map was keeping track of the total amount of gold
in the game (including unmined gold ore, gold at flags/inventories/
other buildings, transported gold/ore, etc.) This counter was
moved to Game and a method on Map was added to return the current
amount of gold in mineral deposits. The name of the method on
Game was changed from get_gold_deposit() to get_gold_total() since
the method is really counting the total gold present in the game,
not just the gold deposited as minerals in the map.