savegame: Throw exception in binary reader on invalid read
Add has_data_left() method on SaveReaderBinary, returning true
if the reader has at least the given number of bytes left.
SaveReaderBinary is changed so that has_data_left() is checked
before every read to guard against reads outside of the underlying
memory area. An exception is thrown when this occurs.