game: Fix bug with initializing player values from save game
Fixes bug where player values were not initialized properly
when loading a game. This adds a call to init() on Player before
loading any save game state. This makes sure that any values
that are not in the save state are initialized correctly. This
fixed a bug where the land ownership value had a random value
after loading a game.