player: Fix init() taking unnecessary game object number
The init() method on Player takes an argument which overwrites
the game object number (index) on the parent class. This is
unnecessary because the game object number is already assigned
by the constructor and should not be changed. This bug became
apparent after 06a0e3b was applied. When certain save games were
loaded the game would freeze on quit.