map-geometry: Remove non-standard directions
Remove DirectionUpRight and DirectionDownLeft from the Direction
enum. These two directions are the non-standard directions
because they do not correspond to a tile edge on the map that
can be traversed. These directions were only used in a few
places for movement and were easily converted to two standard
direction movements.